ev-concierge/requirements.txt
boto3>=1.34.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
# Avoid pyarrow build issues with Python 3.14 by using --no-deps for streamlit if needed
